### 2018-04-26 #706 @stephenturner
possible reviewers: @tgerke @giraola @tomsing1
