### 2018-04-27 f08fe40
fix references, minor reorganization
