### 2018-04-27 #706 @stephenturner
Thanks @arfon. Made a minor change to the paper (fix references, moved a sentence).

Question - is there a way to render the PDF locally as it is rendered here, using pandoc?
