### 2018-04-28 #706 @tgerke
@arfon - I saw the suggestion from @stephenturner for my review of this. Happy to help if needed.
