### 2018-05-10 #706 @leeper
@arfon @mgymrek I can edit this one.
