### 2018-05-10 #731 @whedon
[REVIEW]: qqman: an R package for visualizing GWAS results using Q-Q and manhattan plots
**Submitting author:** @stephenturner (<a href="http://orcid.org/0000-0001-9140-9028">Stephen Turner</a>)
**Repository:** <a href="https://github.com/stephenturner/qqman/" target ="_blank">https://github.com/stephenturner/qqman/</a>
**Version:** v0.1.6
