### 2018-05-10 #731 @leeper
@stephenturner Two quick things:

1. I suspect the paper compilation issue is due to the jpeg image file. Can you convert to PNG perhaps?
2. Please add DOIs for all articles in your .bib file.
