+++ misc/paper/paper.bib
@@ -6,6 +6,7 @@
   number={3},
   pages={559--575},
   year={2007},
+  doi={10.1086/519795},
   publisher={Elsevier}
 }
 
@@ -17,6 +18,7 @@
   number={2},
   pages={263--265},
   year={2004},
+  doi={10.1093/bioinformatics/bth457},
   publisher={Oxford University Press}
 }
 
@@ -28,5 +30,6 @@
   number={18},
   pages={2336--2337},
   year={2010},
+  doi={10.1093/bioinformatics/btq419},
   publisher={Oxford University Press}
 }
