### 2018-05-10 #731 @stephenturner
I think it's because the repository might be pointing to https://github.com/stephenturner/qqman/tree/master/misc/paper instead of to the top level https://github.com/stephenturner/qqman/ ? @leeper
