### 2018-05-10 #731 @stephenturner
See #706 @arfon https://github.com/openjournals/joss-reviews/issues/706#issuecomment-384835018
