### 2018-05-10 #731 @stephenturner
@leeper also, updated .bib with DOI.
