### 2018-05-17 #731 @tgerke
Review complete, all boxes checked. Great package; my lab has already used it multiple times in the past year with success.

Very minor comments (that I don't think should hold back acceptance in JOSS):
-- vignette on CRAN showing zscore as part of gwasResults object before the variable is created later in the vignette.
-- this could be related to CRAN still hosting v0.1.4. The GitHub version does sync with the JOSS paper though (v0.1.6).
-- in our previous use of the package, we sometimes wanted to annotate SNPs in a way that is not related to p-value (e.g. SNPs that were candidates from previous literature). To my knowledge, the package only lets you annotate by p threshold. The feature to highlight by a character vector of SNPs of interest is very useful to this end; the capacity to extend the highlight argument to similarly permit annotation by rsID (or similar) would be helpful in future versions.
