### 2018-05-17 #731 @stephenturner
Thanks for the feedback, @tgerke.

I've never noticed that the vignette on CRAN shows the z-score before the object is created. I'm not sure how that even happened, unless I somehow had the object in my environment when I created the vignette without using devtools. In any case, I've ensured the vignette in 0.1.6 does not do this. I just submitted 0.1.6 to winbuilder. If that comes back okay, and once this is published, I'll add a CITATION file and release 0.1.6 to CRAN.

On the ability to annotate SNPs not related to p-values - I already have some code in place to highlight SNPs based on ID, so it shouldn't be too difficult to use this framework to label them. I've been putting this off because I really want to move this back to ggplot2, to be able to take advantage of the ggrepel package for doing this. That might be version 0.2.0.
