### 2018-05-17 7690a4f
updated news file
