### 2018-05-17 #731 @stephenturner
Looks like it's still failing to generate the PDF because the repo points to https://github.com/stephenturner/qqman/tree/master/misc/paper instead of the top level https://github.com/stephenturner/qqman
