### 2018-05-18 #731 @leeper
@stephenturner Thanks for your package. Everything looks good. Can you please create a DOI for the package using, for example, figshare or zenodo and post the DOI for the archive here?
